Minor update to Scrabble.py
Fixed letter J incorrectly having a score value of 1, and allowed the first word of the game to go anywhere as long as it touches the center star instead of forcing the first letter to be on the star. The center star now also gives double word score, as the official Scrabble rules declare.